karhoo/requirements.txt
flask[async]
mysql-connector-python
requests
redis
pandas
numpy 
